Split struct_rusage into a separate file
This is an attempt to work around the intermittent
mypy crashes while running the test cases.
